requirements.txt
dj-database-url==1.0.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-import-export==3.0.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
gitdb==4.0.9
GitPython==3.1.29
gunicorn==20.1.0
Pillow==9.2.0
requests==2.27.1




